refactor: move send-email-cg.ts and edit-cg.ts actions in contact groups actions
