+++ src/views/contact-groups/actions/edit-cg.ts
@@ -8,9 +8,9 @@ import { useCallback, useMemo } from 'react';
 import { addBoard, getBoardById, reopenBoards, setCurrentBoard } from '@zextras/carbonio-shell-ui';
 import { useTranslation } from 'react-i18next';
 
-import { UIAction } from './types';
-import { ACTION_IDS, EDIT_CONTACT_GROUP_BOARD_ID } from '../constants';
-import { ContactGroup } from '../model/contact-group';
+import { UIAction } from '../../../actions/types';
+import { ACTION_IDS, EDIT_CONTACT_GROUP_BOARD_ID } from '../../../constants';
+import { ContactGroup } from '../../../model/contact-group';
 
 export type EditActionCG = UIAction<ContactGroup, ContactGroup>;
